backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
aiohttp
pyjwt
slowapi
python-jose
uuid6
google-generativeai
numpy
pandas
firebase-admin
